build: make clang builds work; fix -march=native on aarch64
Five build-surface defects. gcc stays the default and no gcc build changes.
All evidence below is from this session on x86_64, g++ 16.2.1 / clang 22.1.8.

* tests/bench_setup.cpp: G.n()/G.m() are node_index/edge_index (unsigned) and
  bench_result stores them as int. gcc pedwarns (-Wnarrowing, 10 instantiations);
  clang stops: "error: non-constant-expression cannot be narrowed from type
  'node_index' (aka 'unsigned int') to 'int' in initializer list" at 121:39.
  Reproduced by recompiling HEAD's copy with clang and the build's own flags.

* benchmarks: rchol's spcol.cpp declares the two gsl_spmatrix_add() scratch
  arrays as VLAs with an initialiser -- clang: "error: variable-sized object may
  not be initialized" (spcol.cpp:44). Patched in the fetched tree next to the
  existing immintrin strip: std::vector of the same length, zero-initialised,
  still in scope for every use of c->i / c->data including the closing sp_copy.
  Gated on Clang deliberately -- a per-call heap allocation is not a stack VLA,
  and gcc must keep timing the pristine upstream code so RCHOL's committed
  numbers cannot move.

* benchmarks: gpu_rchol's experiment/auxilliary.hpp uses INT_MAX with no
  <climits>; that is the copy the CPU driver sees (experiment/custom_cg.hpp
  includes it by quoted name from the same directory). Same fetched-tree patch,
  guarded so re-configure is a no-op. UNVERIFIED: BUILD_GPU_RCHOL needs MKL LP64
  + CUDA and was not built here.

* octave/build.sh: mkoctfile compiles its own SOVERSION stub with $CC under
  -flto=auto, so a gcc $CC under a clang $CXX silently drops
  __octave_mex_soversion__ and the .mex will not load. Derive CC from CXX, refuse
  a mixed pair, and link exactly one OpenMP runtime (-lgomp for gcc, -lomp for
  clang) -- libomp and libgomp in one Octave process segfault it.

* cmake/apxchol_native_arch.cmake -- the real bug: -march=native is the x86
  spelling. clang/aarch64 accepts it and silently falls back to -target-cpu
  generic, so every ARM clang build so far was untuned and any number taken from
  one measured the wrong binary. The module probes with check_cxx_compiler_flag,
  preferring -mcpu=native on ARM and -march=native elsewhere, and emits no flag
  at all if a toolchain accepts neither. Shared by the root project and by
  benchmarks/, which is a separate cmake project and so includes it by relative
  path. octave/build.sh does the same probe in shell.

gcc unmoved, measured: the 31 entries of compile_commands.json are byte-identical
before and after modulo the build-dir name (diff empty); the emitted flag is still
-march=native on this box; ctest 249/249 with an identical test-name list
(md5 cd8848738f0222e04243e2ac6cc14849 both sides); benchmarks' fetched spcol.cpp
still md5 f913a1dedab271a15d6836743c5b4ee0, i.e. untouched.

clang now builds, measured: root project ctest 249/249 with
APXCHOL_BUILD_TOOLS=ON; benchmarks rchol_lib compiles and links; octave MEX
builds and passes 9/9 in Octave 11.3.0 (the .mex needs libomp alone -- libgomp
still arrives transitively via Octave's own liboctmex, in the gcc build too).

+++ tests/bench_setup.cpp
@@ -118,7 +118,11 @@ static bench_result run_one(const char* sname, const char* graph_name,
     double build_ms  = std::chrono::duration<double, std::milli>(t1 - t0).count();
     double factor_ms = std::chrono::duration<double, std::milli>(t2 - t1).count();
 
-    bench_result r{sname, graph_name, G.n(), G.m(),
+    // n()/m() are node_index/edge_index (unsigned); bench_result reports them as
+    // int. Narrowing in a braced init-list is a warning under gcc but an error
+    // under clang, so spell the conversion out.
+    bench_result r{sname, graph_name,
+                   static_cast<int>(G.n()), static_cast<int>(G.m()),
                    static_cast<long long>(F.L.nonZeros()),
                    build_ms, factor_ms, build_ms + factor_ms,
                    F.peak_graph_bytes / (1024.0 * 1024.0)};
